requirement.txt
Flask==1.0.2
Flask-Cors==3.0.7
Flask-MySQLdb==0.2.0
flask-restplus==0.12.1
mysqlclient==1.3.12
requests==2.21.0
pyjwt